+++ src/queries/utils/buildQuery.js
@@ -7,9 +7,11 @@ import {
 
 const buildQuery = ({
   searchIndex,
+  comparator = '==',
+  seeAlsoJoin = 'or',
   isExcludedSeeFromLimiter = false,
 }) => {
-  const indexData = searchableIndexesMap[searchIndex || searchableIndexesValues.KEYWORD];
+  const indexData = searchableIndexesMap[searchIndex];
 
   if (!indexData) {
     return '';
@@ -24,7 +26,7 @@ const buildQuery = ({
   const queryStrings = indexData.map(data => {
     const queryParts = [];
 
-    const queryTemplate = name => `${name}=="%{query}"`;
+    const queryTemplate = name => `${name}${comparator}"%{query}"`;
 
     if (data.plain) {
       const query = queryTemplate(data.name);
@@ -33,6 +35,10 @@ const buildQuery = ({
     }
 
     if (isExcludedSeeFromLimiter) {
+      if (searchIndex === searchableIndexesValues.KEYWORD) {
+        return [`${searchableIndexesValues.KEYWORD}=="%{query}" and authRefType == "Authorized"`];
+      }
+
       return queryParts;
     }
 
@@ -68,7 +74,7 @@ const buildQuery = ({
   });
 
   const flattenedQueryStrings = queryStrings.reduce((acc, arr) => acc.concat(arr));
-  const joinedQueryParts = flattenedQueryStrings.join(' or ');
+  const joinedQueryParts = flattenedQueryStrings.join(` ${seeAlsoJoin} `);
 
   return `(${joinedQueryParts})`;
 };